requirements.txt
Django==4.0
psycopg2-binary==2.8.6
gunicorn==20.0.4
tzdata
pytz
six~=1.14.0
whitenoise==6.4.0
django-dotenv
python-dotenv~=1.0.0
django-environ
